added authorization on row actions in table view

+++ Master_v2.Server/Controllers/ArticlesController.cs
@@ -15,13 +15,21 @@ namespace Master_v2.Server.Controllers
         [HttpGet]
 		[Route("api/articles")]
 		public IEnumerable<Article> Get(){
+            if (!AuthorizationStore.checkReadPermission("article"))
+            {
+                return new List<Article>();
+            }
 			return _repository.GetAll();
 		}
 
 		[HttpGet]
 		[Route("api/article/{id}")]
 		public Article GetById(int id){
-			if(id==0){
+            if (!AuthorizationStore.checkReadPermission("article"))
+            {
+                return new Article();
+            }
+            if (id==0){
 				return new Article();
 			}
 			else{
@@ -32,21 +40,33 @@ namespace Master_v2.Server.Controllers
 		[HttpPost]
 		[Route("api/article/create")]
 		public void Post([FromBody] Article model){
-			if (ModelState.IsValid){
+            if (!AuthorizationStore.checkWritePermissions("article"))
+            {
+                return;
+            }
+            if (ModelState.IsValid){
 				_repository.Add(model);
 			}
 		}
 		[HttpPost]
 		[Route("api/article/edit")]
 		public void Update([FromBody] Article model){
-			if (ModelState.IsValid) {
+            if (!AuthorizationStore.checkWritePermissions("article"))
+            {
+                return;
+            }
+            if (ModelState.IsValid) {
 				_repository.Update(model);
 			}
 		}
 		[HttpDelete]
 		[Route("api/article/delete/{id}")]
 		public void Delete(int id){
-			_repository.Delete(id);
+            if (!AuthorizationStore.checkWritePermissions("article"))
+            {
+                return ;
+            }
+            _repository.Delete(id);
 		}
 	}
 }
